Refactor code for getting dynamic movie information.
